refactor: split pos product catalog route
